+++ src/common/Utilities.py
@@ -9,8 +9,8 @@ import sirf.pyiutilities as pyiutil
 import re
 
 __licence__ = """SyneRBI Synergistic Image Reconstruction Framework (SIRF)
-Copyright 2015 - 2019 Rutherford Appleton Laboratory STFC
-Copyright 2015 - 2020 University College London
+Copyright 2015 - 2021 Rutherford Appleton Laboratory STFC
+Copyright 2015 - 2021 University College London
 
 This is software developed for the Collaborative Computational
 Project in Synergistic Reconstruction for Biomedical Imaging (formerly CCP PETMR)
@@ -67,7 +67,7 @@ def existing_filepath(data_path, file_name):
     data_path: path to the file
     file_name: file name
     '''
-    full_name = data_path + '/' + file_name
+    full_name = os.path.join(os.path.abspath(data_path), file_name)
     if not os.path.isfile(full_name):
         raise error('file %s not found' % full_name)
     return full_name
@@ -544,3 +544,200 @@ def is_operator_adjoint(operator, num_tests = 5, max_err = 10e-5, verbose = True
             elif verbose:
                 print("Pass, with a with normalized error of " + str(norm_err) + " (max: " + str(max_err) + ")")
     return True
+
+
+class TestDataContainerAlgebra(object):
+
+    '''A base class for unit test of DataContainer algebra.'''
+    def test_divide_scalar(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image1.fill(1.)
+        image2.fill(2.)
+        
+        tmp = image1/1.
+        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
+    
+        tmp1 = image1.divide(1.)
+        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
+        
+        image1.divide(1., out=image2)
+        numpy.testing.assert_array_equal(tmp.as_array(), image2.as_array())
+
+    def test_divide_datacontainer(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        # add 1 because the data contains zeros and divide is not going to be happy
+        image1 = self.image1 + 1
+        image2 = self.image2 + 1
+        
+        tmp = image1/image2
+
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp.as_array()
+            )
+    
+        tmp1 = image1.divide(image2)
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+        
+        tmp1.fill(2.)
+        image1.divide(image2, out=tmp1)
+        
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+
+    def test_multiply_scalar(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image2.fill(2.)
+        
+        tmp = image1 * 1.
+        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
+    
+        tmp1 = image1.multiply(1.)
+        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
+        
+        image1.multiply(1., out=image2)
+        numpy.testing.assert_array_equal(tmp.as_array(), image2.as_array())
+
+    def test_multiply_datacontainer(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image2.fill(1.)
+        tmp = image1 * image2
+
+        numpy.testing.assert_array_almost_equal(
+            image1.as_array(), tmp.as_array()
+            )
+    
+        tmp1 = image1.multiply(image2)
+        numpy.testing.assert_array_almost_equal(
+            image1.as_array(), tmp1.as_array()
+            )
+        
+        tmp1.fill(2.)
+        image1.multiply(image2, out=tmp1)
+        
+        numpy.testing.assert_array_almost_equal(
+            image1.as_array(), tmp1.as_array()
+            )
+
+    def test_add_scalar(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image1.fill(0)
+        image2.fill(1)
+        
+        tmp = image1 + 1.
+        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
+    
+        tmp1 = image1.add(1.)
+        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
+        
+        tmp1.fill(0)
+        image1.add(1., out=tmp1)
+        numpy.testing.assert_array_equal(tmp1.as_array(), image2.as_array())
+    
+    def test_add_datacontainer(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image1.fill(0.)
+        image2.fill(1.)
+        tmp = image1 + image2
+
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp.as_array()
+            )
+    
+        tmp1 = image1.add(image2)
+        
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+        
+        tmp1.fill(2.)
+        image1.add(image2, out=tmp1)
+        
+        numpy.testing.assert_array_almost_equal(
+            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+        
+    
+    def test_subtract_scalar(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        image1.fill(2)
+        image2.fill(1)
+        
+        tmp = image1 - 1.
+        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
+    
+        tmp1 = image1.subtract(1.)
+        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
+        
+        tmp1.fill(0)
+        image1.subtract(1., out=tmp1)
+        numpy.testing.assert_array_equal(tmp1.as_array(), image2.as_array())
+
+    def test_subtract_datacontainer(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        image1 = self.image1
+        image2 = self.image2
+        
+        tmp = image1 - image2
+
+        numpy.testing.assert_array_almost_equal(
+            numpy.zeros(image1.shape, dtype=numpy.float32), tmp.as_array()
+            )
+    
+        tmp1 = image1.subtract(image2)
+        
+        numpy.testing.assert_array_almost_equal(
+            numpy.zeros(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+        
+        tmp1.fill(2.)
+        image1.subtract(image2, out=tmp1)
+        
+        numpy.testing.assert_array_almost_equal(
+            numpy.zeros(image1.shape, dtype=numpy.float32), tmp1.as_array()
+            )
+
+    def test_division_by_scalar_zero(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        try:
+            self.image1 / 0.
+            self.assertFalse(True)
+        except ZeroDivisionError:
+            self.assertTrue(True)
+        except error:
+            self.assertTrue(True)
+    
+    def test_division_by_datacontainer_zero(self):
+        if hasattr(self, 'cwd'):
+            os.chdir(self.cwd)
+        try:
+            self.image2 *= 0
+            tmp = self.image1 / self.image2
+            self.assertFalse(True)
+        except ZeroDivisionError:
+            self.assertTrue(True)
+        except error:
+            self.assertTrue(True)